adding refresh token to existing /api/auth route
